requirements.txt
asgiref==3.4.1
cffi==1.14.6
cryptography==35.0.0
Django==3.2.8
django-rest-knox==4.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.3.2
psycopg2-binary==2.9.1
pycparser==2.20
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
